% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{gendata}
\alias{gendata}
\title{Simulates data from a Dirichlet-Multinomial regression model.}
\usage{
gendata(
  J = 10,
  n = 100,
  P = 10,
  Q = 5,
  beta_min = 0.75,
  beta_max = 1.5,
  n_rel_taxa = 2,
  n_rel_x = 2,
  n_rel_z = 2,
  n_rel_inter = c(2, 2, 3, 3),
  corrx = 0.4,
  corrz = 0.4,
  theta0 = 0.01,
  WH = FALSE,
  nli = TRUE,
  randint = FALSE,
  repmeas = 3
)
}
\arguments{
\item{J}{number of taxa}

\item{n}{number of samples}

\item{P}{number of continuous covariates}

\item{Q}{number of discrete covariates}

\item{beta_min}{minimum absolute value of the regression parameters}

\item{beta_max}{maximum absolute value of the regression parameters}

\item{n_rel_taxa}{number of relevant taxa}

\item{n_rel_x}{number of relevant continuous covariates}

\item{n_rel_z}{number of relevant discrete covariates}

\item{n_rel_inter}{vector of length four
\enumerate{
\item number of continuous-discrete interactions
\item number of discrete-discrete interactions
\item number of linear continuous-continuous interactions
\item number of nonlinear continuous-continuous interactions
}}

\item{corrx}{the correlation between continuous covariates}

\item{corrz}{the correlation between discrete covariates}

\item{theta0}{the overdispersion parameter}

\item{WH}{the weak heredity flag}

\item{nli}{whether to include nonlinear interactions between continuous covariates}

\item{randint}{whether should be included a random intercept}

\item{repmeas}{number of repeated measurements for each sample}
}
\value{
Y: (count matrix) \eqn{n \times J}

X: (design matrix) \eqn{n \times P} (continuous variables)

Z: (design matrix) \eqn{n \times Q} (discrete variables)

intercept: simulated intercept \eqn{J-}dimensional vector

thetax: simulated coefficient matrix \eqn{P \times J}

thetaz: simulated coefficient matrix \eqn{Q \times J}

inter_xz: simulated coefficient matrix (continuous-discrete
interactions) \eqn{P \times Q \times J}

inter_xx: simulated coefficient matrix (continuous-continuous linear interactions)
\eqn{P \times P \times J}

inter_xx_nl: simulated coefficient matrix (continuous-continuous nonlinear interactions)
\eqn{n \times P \times J}

nlind: indicator for nonlinear interactions

inter_zz: simulated coefficient matrix (discrete-discrete
interactions) \eqn{Q \times Q \times J}

RI: simulated random intercept \eqn{n-}dimensional vector

GI: group indicator for repeated measurements

ZETA: log linear predictor

part1, part2: subject-specific coefficients
}
\description{
\code{gendata} returns the data simulated according to input parameters
}
\examples{
gendata(10, 100, 10, 5, 0.75, 1.5, 2, 2, 2, c(2, 2, 3, 3), 0.4, 0.4, 0.01, FALSE, TRUE)
}
