% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapfn.R
\name{ssdm}
\alias{ssdm}
\title{an R wrapper to Rcpp function sampler}
\usage{
ssdm(
  YY,
  XX,
  ZZ,
  Niter,
  burn,
  thin,
  upsv = FALSE,
  hereditariety = 2,
  conlnl = TRUE,
  randint = TRUE,
  hyper_theta_x = c(0.05, 10),
  hyper_theta_z = c(0.05, 10),
  penmig_lin = c(5, 25, 0.00025, 1, 1),
  penmig_nl = c(5, 25, 0.00025, 1, 1),
  tx = c(0, 1),
  tz = c(0, 1),
  prior_int = c(0, 10),
  init = TRUE,
  init_fd = 0.1
)
}
\arguments{
\item{YY}{(count matrix) rows n samples, columns: J species}

\item{XX}{(design matrix) n x P (continuous variables)}

\item{ZZ}{(design matrix) n * Q (discrete variables)}

\item{Niter}{number of MCMC iteration}

\item{burn}{number of MCMC iteration discarded due to burn-in period}

\item{thin}{thinning, retain one iteration each \code{thin}}

\item{upsv}{flag to update slab variance (default is \code{FALSE})}

\item{hereditariety}{hereditariety assumption on interactions. If \code{2}
strong heredity is assumed, if \code{1} weak heredity is assumed, if
\code{0} no assumption on heredity condition is made.
Default is \code{TRUE}}

\item{conlnl}{it is a constrain imposed on nonlinear continuous continuous
interactions. Default is \code{TRUE}. A nonlinear continuous continuous interaction
may be included if and only if the correspondent linear is included.}

\item{randint}{random intercept is considered for different observation from
the same patient. Default is \code{TRUE} (per ora FALSE)}

\item{hyper_theta_x}{vector of dimension two. The first element is the a
parameter of Beta function after eq 12 manuscript and default is
\code{a = 0.02}. The second element is the variance of the slab. It is not
updated and default is \code{tau_slab = 10}.}

\item{hyper_theta_z}{vector of dimension two. The first element is the a
parameter of Beta function after eq 12 manuscript and default is
\code{a = 0.2}. The second element is the variance of the slab. It is not
updated and default is \code{tau_slab = 10}.}

\item{penmig_lin}{vector of length 5.
These are the hyperparameters for
peNMIG on linear continuous continuous interactions. The first two are
\code{a_tau, b_tau} that are the parameters for Inverse Gamma on
\code{alpha}'s variance. The third is \code{v_0}. The remaining are
\code{a_omega, b_omega}.

Default is \code{penmig_lin = c(5.0, 25.0, 0.0025, 1.0, 1.0)}}

\item{penmig_nl}{vector of length 5.
These are the hyperparameters for
peNMIG on nonlinear continuous continuous interactions. The first two are
\code{a_tau, b_tau} that are the parameters for Inverse Gamma on
\code{alpha}'s variance. The third is \code{v_0}. The remaining are
\code{a_omega, b_omega}.

Default is \code{penmig_lin = c(5.0, 25.0, 0.0025, 1.0, 1.0)}}

\item{tx}{vector of length 2. Hyperparameters for Uniform prior on
first threshold function's parameter. Default is \code{tx = c(0, 0.25)}}

\item{tz}{vector of length 2. Hyperparameters for Uniform prior on
second threshold function's parameter. Default is \code{tz = c(0, 0.25)}}

\item{prior_int}{vector of length 2. Hyperparameters for Normal prior on model's intercept.
Default is \code{prior_int = c(0, 10)}}

\item{init}{whether main effects coefficients shold be initialized with marginal correlations. Default is \code{TRUE}}

\item{init_fd}{false discovery rate adopted as threshold if \code{init = TRUE}}
}
\value{
thetaxposterior
samples from posterior distribution of \eqn{ \theta_x }

thetazposterior samples from posterior distribution of \eqn{ \theta_z }

muposterior samples from posterior distribution of \eqn{ \mu }

thrxposterior samples from posterior distribution of \eqn{t_x}

thrzposterior samples from posterior distribution of \eqn{t_z}

loglinpred log linear predictor at each MCMC iteration

loglik loglikelihood at each MCMC iteration

alpha0posterior samples from posterior distribution of \eqn{\alpha^0}

alphastarposterior samples from posterior distribution of \eqn{\alpha^{\star}}

eta0posterior samples from posterior distribution of \eqn{\eta^{0}}

etastarposterior samples from posterior distribution of \eqn{\eta^{\star}}

xi0posterior samples from posterior distribution of \eqn{\dot{\xi}^{0}}

xistarposterior samples from posterior distribution of \eqn{\dot{\xi}^{\star}}

bxzposterior samples from posterior distribution of \eqn{b^{xz}}

bzzposterior samples from posterior distribution of \eqn{b^{zz}}

betaxz samples from posterior distribution of \eqn{\beta(x,z)}

betaz samples from posterior distribution of \eqn{\beta(z)}

rint samples from posterior distribution of \eqn{\iota}

varrint samples from posterior distribution of \eqn{\sigma_{\iota}}
}
\description{
MCMC sampler to perform bayesian variable selection, hierarchical Dirichlet-Multinomial regression
}
