% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdmreg.R
\docType{package}
\name{ssdmreg}
\alias{ssdmreg}
\title{Subject Specific Dirichlet Multinomial Regression from Microbiota analysis}
\description{
The implementation has been done in \code{C++} through the use of \code{Rcpp} and \code{RcppArmadillo}.
}
\author{
Matteo Pedone
}
